MEGA REFACTOR de estructura
