+++ api/notes/toDo.txt
@@ -1,0 +1,3 @@
+
+- Modularizar mejor nodemailer
+- Hacer emails según idioma
