+++ api/src/routes/subscription/subscription-routes.ts
@@ -1,35 +1,40 @@
 import { Router } from "express";
 import jwtCheck from "../../config/jwtMiddleware";
-import {
-  handleFindAllSubscriptionsRequest,
-  handleCreateNewSubscriptionRequest,
-  handleDeleteSubscriptionByIdRequest,
-  handleUpdateSubscriptionByIdRequest,
-  handleGetUserSubscriptionsRequest,
-} from "./subscription-middlewares";
+import subscriptionControllers from "../../controllers/subscription";
 
 const router = Router();
 
-router.get("/findAll", handleFindAllSubscriptionsRequest);
+router.get(
+  "/findAll",
+  subscriptionControllers.handleFindAllSubscriptionsRequest
+);
 
 // CREATE NEW SUBSCRIPTION :
-router.post("/", jwtCheck, handleCreateNewSubscriptionRequest);
+router.post(
+  "/",
+  jwtCheck,
+  subscriptionControllers.handleCreateNewSubscriptionRequest
+);
 
 // DELETE SUBSCRIPTION :
 router.delete(
   "/:subscription_id",
   jwtCheck,
-  handleDeleteSubscriptionByIdRequest
+  subscriptionControllers.handleDeleteSubscriptionByIdRequest
 );
 
 // UPDATE SUBSCRIPTION :
 router.patch(
   "/:subscription_id",
   jwtCheck,
-  handleUpdateSubscriptionByIdRequest
+  subscriptionControllers.handleUpdateSubscriptionByIdRequest
 );
 
 // GET ALL USER SUBSCRIPTIONS :
-router.get("/userSubs", jwtCheck, handleGetUserSubscriptionsRequest);
+router.get(
+  "/userSubs",
+  jwtCheck,
+  subscriptionControllers.handleGetUserSubscriptionsRequest
+);
 
 export default router;
