+++ api/src/routes/user/user-routes.ts
@@ -1,32 +1,29 @@
 import { Router } from "express";
 import jwtCheck from "../../config/jwtMiddleware";
-import {
-  handleFindAllUsersRequest,
-  handleRegisterNewUserRequest,
-  handleUserExistsInDBRequest,
-  handleDeleteAllUserDataRequest,
-  handleGetUserInfoRequest,
-  handleUpdateUserSanitizingRequest,
-} from "./user-middlewares";
+import userControllers from "../../controllers/user";
 
 const router = Router();
 
 // FIND ALL USERS (JUST FOR TESTING. NOT FOR PRODUCTION) :
-router.get("/findAll", handleFindAllUsersRequest);
+router.get("/findAll", userControllers.findAllUsersHandler);
 
 // GET USER INFO :
-router.get("/userInfo", jwtCheck, handleGetUserInfoRequest);
+router.get("/userInfo", jwtCheck, userControllers.getUserInfoHandler);
 
 // CREATE/REGISTER NEW USER :
-router.post("/register", jwtCheck, handleRegisterNewUserRequest);
+router.post("/register", jwtCheck, userControllers.registerNewUserHandler);
 
 // USER EXSITS IN DB. res = {msg: true / false}
-router.get("/existsInDB", jwtCheck, handleUserExistsInDBRequest);
+router.get("/existsInDB", jwtCheck, userControllers.userExistsInDBHandler);
 
 // UPDATE USER INFO :
-router.patch("/update", jwtCheck, handleUpdateUserSanitizingRequest);
+router.patch("/update", jwtCheck, userControllers.updateUserSanitizingHandler);
 
 // DELETE ALL USER DATA FROM DB :
-router.delete("/destroyAll/:_id", jwtCheck, handleDeleteAllUserDataRequest);
+router.delete(
+  "/destroyAll/:_id",
+  jwtCheck,
+  userControllers.deleteAllUserDataHandler
+);
 
 export default router;
